style: refine 명칭을 sanitize 로 일괄 변경
