+++ packages/util/sanitizePageNumber.ts
@@ -1,0 +1,19 @@
+/**
+ * 페이지 번호를 정제한다.
+ *
+ * 0보다 큰 양의 정수만 통과시킨다. (문자열 입력 가능)
+ *
+ * const result = sanitizePageNumber('-1');
+ *
+ * console.log(result); // 1
+ * @param raw
+ * @returns
+ */
+export function sanitizePageNumber(raw: unknown) {
+  const page = Math.floor(Math.abs(Number(raw || 1)));
+
+  if (Number.isFinite(page) && page > 0) {
+    return page;
+  }
+  return 1;
+}
